workable
I dropped the 1st fallback for when there's no poco-designated contact. It was going to be to some contact (last listed) that has contact.contactTypeIdValue == 'Notified', that was inherited from the vendor. I should be able to check the vendor inheritance using fromPartyType = 'vendor', but somehow I can't manage to do this check without crashing the notification.

So now it's just: either a designated contact, or nothing.